+++ example/urls.py
@@ -13,6 +13,8 @@ urlpatterns = patterns('',
 
     url(r'^column/sequence/$', 'app.views.sequence_column', name='sequence_column'),
     url(r'^column/calendar/$', 'app.views.calendar_column', name='calendar_column'),
+    url(r'^column/link/$', 'app.views.link_column', name='link_column'),
 
+    url(r'^user/(\d+)/$', 'app.views.user_profile', name='user_profile'),
     url(r'^table/', include('table.urls')),
 )