+++ CHANGELOG.md
@@ -12,7 +12,7 @@
     * space
 
 * Refactoring table module
-* Inherite columns from parent table class
+* Support subclass table
 
 ## v0.1.1